% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ambil_baris}
\alias{ambil_baris}
\title{Ambil Baris}
\usage{
ambil_baris(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Posisi baris}
}
\value{
Dataset yang diambil barisnya
}
\description{
Wrapper untuk dplyr::slice()
}
