% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_wrapper.R
\name{blntglthn}
\alias{blntglthn}
\title{blntglthn}
\usage{
blntglthn(x)
}
\arguments{
\item{x}{Karakter yang berisi tanggal dalam format tanggal-bulan-tahun, misal "27-10-2025"}
}
\value{
Objek Date
}
\description{
Mengubah tanggal dari format "DD-MM-YYYY" menjadi objek Date.
}
