% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{buat_kolom_saja}
\alias{buat_kolom_saja}
\title{Buat Kolom Saja}
\usage{
buat_kolom_saja(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Kolom baru}
}
\value{
Dataset hanya berisi kolom baru
}
\description{
Wrapper untuk dplyr::transmute()
}
