% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_jumlah}
\alias{dasaR_jumlah}
\title{Menjumlahkan nilai vektor}
\usage{
dasaR_jumlah(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan untuk `sum()`}
}
\value{
Sebuah nilai numerik hasil penjumlahan
}
\description{
Padanan fungsi `sum()`
}
\examples{
dasaR_jumlah(c(1, 2, 3))  # 6

}
\seealso{
[base::sum()]

Other dasaR: 
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
