% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_maksimum}
\alias{dasaR_maksimum}
\title{Mengambil nilai maksimum}
\usage{
dasaR_maksimum(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan}
}
\value{
Nilai maksimum
}
\description{
Padanan fungsi `max()`
}
\examples{
dasaR_maksimum(c(3, 1, 6))  # 6

}
\seealso{
[base::max()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
