% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_median}
\alias{dasaR_median}
\title{Menghitung nilai median}
\usage{
dasaR_median(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan untuk `median()`}
}
\value{
Nilai median
}
\description{
Padanan fungsi `median()`
}
\examples{
dasaR_median(c(10, 2, 5))  # 5

}
\seealso{
[stats::median()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
