% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_minimum}
\alias{dasaR_minimum}
\title{Mengambil nilai minimum}
\usage{
dasaR_minimum(x, ...)
}
\arguments{
\item{x}{Vektor numerik.}

\item{...}{Argumen tambahan untuk `min()`}
}
\value{
Nilai minimum
}
\description{
Padanan fungsi `min()`
}
\examples{
dasaR_minimum(c(3, 1, 6))  # 1

}
\seealso{
[base::min()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
