% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_panjang}
\alias{dasaR_panjang}
\title{Menghitung panjang vektor}
\usage{
dasaR_panjang(x)
}
\arguments{
\item{x}{Objek R (biasanya vektor)}
}
\value{
Bilangan bulat panjang objek
}
\description{
Padanan fungsi `length()`
}
\examples{
dasaR_panjang(c(10, 20, 30))  # 3

}
\seealso{
[base::length()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
