% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_rata_rata}
\alias{dasaR_rata_rata}
\title{Menghitung nilai rata-rata}
\usage{
dasaR_rata_rata(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan untuk `mean()`}
}
\value{
Nilai rata-rata
}
\description{
Padanan fungsi `mean()`
}
\examples{
dasaR_rata_rata(c(1, 2, 3))  # 2

}
\seealso{
[base::mean()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
