% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_rentang}
\alias{dasaR_rentang}
\title{Menghitung nilai rentang}
\usage{
dasaR_rentang(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan}
}
\value{
Vektor panjang 2 berisi nilai minimum dan maksimum
}
\description{
Padanan fungsi `range()`
}
\examples{
dasaR_rentang(c(3, 1, 6))  # 1 6

}
\seealso{
[base::range()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
