% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_standar_deviasi}
\alias{dasaR_standar_deviasi}
\title{Menghitung standar deviasi}
\usage{
dasaR_standar_deviasi(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan}
}
\value{
Nilai standar deviasi
}
\description{
Padanan fungsi `sd()`
}
\examples{
dasaR_standar_deviasi(c(1, 2, 3))

}
\seealso{
[stats::sd()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_unik}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
