% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_unik}
\alias{dasaR_unik}
\title{Mengambil nilai unik}
\usage{
dasaR_unik(x)
}
\arguments{
\item{x}{Vektor}
}
\value{
Vektor berisi nilai unik
}
\description{
Padanan fungsi `unique()`
}
\examples{
dasaR_unik(c(1, 1, 2, 3))

}
\seealso{
[base::unique()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_variansi}()}
}
\concept{dasaR}
