% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{dasaR_variansi}
\alias{dasaR_variansi}
\title{Menghitung variansi}
\usage{
dasaR_variansi(x, ...)
}
\arguments{
\item{x}{Vektor numerik}

\item{...}{Argumen tambahan}
}
\value{
Nilai variansi
}
\description{
Padanan fungsi `var()`
}
\examples{
dasaR_variansi(c(1, 2, 3))

}
\seealso{
[stats::var()]

Other dasaR: 
\code{\link{dasaR_jumlah}()},
\code{\link{dasaR_maksimum}()},
\code{\link{dasaR_median}()},
\code{\link{dasaR_minimum}()},
\code{\link{dasaR_panjang}()},
\code{\link{dasaR_rata_rata}()},
\code{\link{dasaR_rentang}()},
\code{\link{dasaR_standar_deviasi}()},
\code{\link{dasaR_unik}()}
}
\concept{dasaR}
