% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_direktori_lihat}
\alias{data_direktori_lihat}
\title{Menampilkan direktori kerja aktif}
\usage{
data_direktori_lihat()
}
\value{
String path direktori kerja
}
\description{
Padanan fungsi `getwd()`
}
\examples{
data_direktori_lihat()
}
\seealso{
[getwd()]
}
