% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_direktori_simpan}
\alias{data_direktori_simpan}
\title{Mengubah direktori kerja aktif}
\usage{
data_direktori_simpan(path)
}
\arguments{
\item{path}{String path direktori baru}
}
\value{
Path baru setelah diubah
}
\description{
Padanan fungsi `setwd()`
}
\examples{
\dontrun{
data_direktori_simpan("C:/Users/Username/Documents/Proyek")
}
}
\seealso{
[setwd()], [getwd()], [direktori_lihat()]
}
