% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_lihat_dimensi}
\alias{data_lihat_dimensi}
\title{Melihat dimensi data (jumlah baris dan jumlah kolom)}
\usage{
data_lihat_dimensi(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Vektor numerik: jumlah baris dan kolom
}
\description{
Padanan fungsi `dim()`
}
\examples{
data_lihat_dimensi(penguin)
}
\seealso{
[base::dim()]
}
