% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_lihat_ekor}
\alias{data_lihat_ekor}
\title{Menampilkan beberapa baris terakhir dari data frame}
\usage{
data_lihat_ekor(data, n = 6)
}
\arguments{
\item{data}{Data frame atau vektor}

\item{n}{Jumlah baris yang ingin ditampilkan (default = 6)}
}
\value{
Data frame berisi baris akhir
}
\description{
Padanan fungsi `utils::tail()`
}
\examples{
data_lihat_ekor(penguin, 3)
}
\seealso{
[utils::tail()]
}
