% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_lihat_kepala}
\alias{data_lihat_kepala}
\title{Menampilkan beberapa baris pertama dari data frame}
\usage{
data_lihat_kepala(data, n = 6)
}
\arguments{
\item{data}{Data frame atau vektor}

\item{n}{Jumlah baris yang ingin ditampilkan (default = 6).}
}
\value{
Data frame berisi baris awal.
}
\description{
Padanan fungsi `utils::head()`
}
\examples{
data_lihat_kepala(penguin, 3)
}
\seealso{
[utils::head()]
}
