% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_lihat_nama}
\alias{data_lihat_nama}
\title{Melihat nama kolom}
\usage{
data_lihat_nama(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Vektor karakter berisi nama kolom
}
\description{
Padanan fungsi `names()`
}
\examples{
data_lihat_nama(penguin)
}
\seealso{
[base::names()]
}
