% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{data_lihat_sekilas}
\alias{data_lihat_sekilas}
\title{Melihat sekilas struktur data}
\usage{
data_lihat_sekilas(data)
}
\arguments{
\item{data}{Data frame yang ingin dilihat strukturnya.}
}
\value{
Invisible NULL; hanya menampilkan output di konsol
}
\description{
Padanan fungsi `dplyr::glimpse()`
}
\examples{
data_lihat_sekilas(penguin)
}
\seealso{
[dplyr::glimpse()]
}
