% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{estetika}
\alias{estetika}
\title{Menentukan pemetaan estetika visualisasi}
\usage{
estetika(...)
}
\arguments{
\item{...}{Pemetaan estetika seperti `x`, `y`, `color`, `shape`, dan `size`}
}
\value{
Objek pemetaan estetika (`aes`)
}
\description{
Padanan fungsi `ggplot2::aes()`
}
\examples{
estetika(x = mpg, y = hp)

}
\seealso{
[ggplot2::aes()]

Other visualisasi: 
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
