% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ganti_nama}
\alias{ganti_nama}
\title{Ganti Nama}
\usage{
ganti_nama(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Kolom baru = kolom lama}
}
\value{
Dataset dengan kolom diubah namanya
}
\description{
Wrapper untuk dplyr::rename()
}
