% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{geometrik_batang}
\alias{geometrik_batang}
\title{Membuat diagram batang (bar chart)}
\usage{
geometrik_batang(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika untuk layer ini.}

\item{...}{Argumen tambahan untuk `ggplot2::geom_bar()`.}
}
\value{
Layer diagram batang untuk ditambahkan ke `ggplot`.
}
\description{
Menampilkan jumlah atau proporsi kategori dalam bentuk batang.
Fungsi ini adalah padanan dari `ggplot2::geom_bar()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = species)) + geometrik_batang()

}
\seealso{
[ggplot2::geom_bar()]
}
