% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{geometrik_garis_tren}
\alias{geometrik_garis_tren}
\title{Menambahkan garis tren atau model ke grafik}
\usage{
geometrik_garis_tren(mapping = NULL, method = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika tambahan untuk layer ini (opsional).}

\item{method}{Metode pemodelan, misalnya `"lm"` untuk regresi linier.}

\item{...}{Argumen tambahan untuk `ggplot2::geom_smooth()`.}
}
\value{
Layer garis tren yang dapat ditambahkan ke objek `ggplot`.
}
\description{
Menampilkan garis tren hasil pemodelan data (seperti regresi linier).
Fungsi ini adalah padanan dari `ggplot2::geom_smooth()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = flipper_length_mm, y = body_mass_g)) +
#   geometrik_garis_tren(method = "lm")

}
\seealso{
[ggplot2::geom_smooth()]
}
