% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{geometrik_histogram}
\alias{geometrik_histogram}
\title{Membuat histogram}
\usage{
geometrik_histogram(mapping = NULL, bins = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika untuk layer ini.}

\item{bins}{Jumlah batang (default mengikuti `ggplot2`).}

\item{...}{Argumen tambahan untuk `ggplot2::geom_histogram()`.}
}
\value{
Layer histogram untuk ditambahkan ke `ggplot`.
}
\description{
Menampilkan distribusi data numerik dalam bentuk batang berkelompok.
Fungsi ini adalah padanan dari `ggplot2::geom_histogram()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = body_mass_g)) + geometrik_histogram()

}
\seealso{
[ggplot2::geom_histogram()]
}
