% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{geometrik_kerapatan}
\alias{geometrik_kerapatan}
\title{Membuat kurva kerapatan (density plot)}
\usage{
geometrik_kerapatan(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika untuk layer ini.}

\item{...}{Argumen tambahan untuk `ggplot2::geom_density()`.}
}
\value{
Layer kurva kerapatan untuk ditambahkan ke `ggplot`.
}
\description{
Menampilkan distribusi probabilitas dari data numerik.
Fungsi ini adalah padanan dari `ggplot2::geom_density()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = body_mass_g)) + geometrik_kerapatan()

}
\seealso{
[ggplot2::geom_density()]
}
