% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{geometrik_plot_kotak}
\alias{geometrik_plot_kotak}
\title{Membuat plotkotak (boxplot)}
\usage{
geometrik_plot_kotak(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika untuk layer ini.}

\item{...}{Argumen tambahan untuk `ggplot2::geom_boxplot()`.}
}
\value{
Layer boxplot untuk ditambahkan ke `ggplot`.
}
\description{
Menampilkan ringkasan statistik distribusi data numerik berdasarkan kategori.
Fungsi ini adalah padanan dari `ggplot2::geom_boxplot()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = species, y = body_mass_g)) + geometrik_kotak()

}
\seealso{
[ggplot2::geom_boxplot()]
}
