% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{geometrik_titik}
\alias{geometrik_titik}
\title{Menambahkan layer titik (scatter plot)}
\usage{
geometrik_titik(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika tambahan untuk layer ini (opsional).}

\item{...}{Argumen tambahan untuk `ggplot2::geom_point()`.}
}
\value{
Layer titik yang dapat ditambahkan ke objek `ggplot`.
}
\description{
Membuat visualisasi titik untuk dua variabel numerik.
Fungsi ini adalah padanan dari `ggplot2::geom_point()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = flipper_length_mm, y = body_mass_g)) +
#   geometrik_titik(estetika(color = species))

}
\seealso{
[ggplot2::geom_point()]
}
