% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafik.R
\name{grafik}
\alias{grafik}
\title{Grafik}
\usage{
grafik(data, mapping = NULL)
}
\arguments{
\item{data}{Dataset (data frame atau tibble)}

\item{mapping}{Aesthetic mapping, gunakan ggplot2::aes()}
}
\value{
Objek ggplot2
}
\description{
Fungsi pembungkus sederhana untuk ggplot2::ggplot()
}
