% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_facet_baris_kolom}
\alias{grafik_facet_baris_kolom}
\title{Facet baris dan kolom}
\usage{
grafik_facet_baris_kolom(rows = NULL, cols = NULL, ...)
}
\arguments{
\item{rows}{Variabel baris}

\item{cols}{Variabel kolom}

\item{...}{Argumen tambahan}
}
\value{
Layer grid facet
}
\description{
Padanan fungsi `ggplot2::facet_grid()`
}
\examples{
visualisasi_ggplot(penguin, estetika(mpg, hp)) +
  grafik_geometrik_titik() +
  grafik_facet_baris_kolom(
    rows = ggplot2::vars(cyl),
    cols = ggplot2::vars(gear)
  )

}
\seealso{
[ggplot2::facet_grid()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
