% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_facet_kolom}
\alias{grafik_facet_kolom}
\title{Facet dengan kolom}
\usage{
grafik_facet_kolom(facets, ...)
}
\arguments{
\item{facets}{Formula untuk pemisahan}

\item{...}{Argumen tambahan}
}
\value{
Layer facet
}
\description{
Padanan fungsi `ggplot2::facet_wrap()`
}
\examples{
visualisasi_ggplot(penguin, estetika(mpg, hp)) +
  grafik_geometrik_titik() +
  grafik_facet_kolom(~ cyl)

}
\seealso{
[ggplot2::facet_wrap()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
