% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_geometrik_densitas}
\alias{grafik_geometrik_densitas}
\title{Membuat plot densitas (density plot)}
\usage{
grafik_geometrik_densitas(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika}

\item{...}{Argumen tambahan}
}
\value{
Layer plot densitas
}
\description{
Padanan fungsi `ggplot2::geom_density()`
}
\examples{
visualisasi_ggplot(penguin, estetika(x = mpg)) + grafik_geometrik_densitas()

}
\seealso{
[ggplot2::geom_density()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
