% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_geometrik_histogram}
\alias{grafik_geometrik_histogram}
\title{Membuat histogram}
\usage{
grafik_geometrik_histogram(mapping = NULL, bins = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika}

\item{bins}{Jumlah batang histogram}

\item{...}{Argumen tambahan}
}
\value{
Layer histogram
}
\description{
Padanan fungsi `ggplot2::geom_histogram()`
}
\examples{
visualisasi_ggplot(penguin, estetika(x = mpg)) + grafik_geometrik_histogram(bins = 10)

}
\seealso{
[ggplot2::geom_histogram()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
