% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_geometrik_titik}
\alias{grafik_geometrik_titik}
\title{Menambahkan layer titik (scatter plot)}
\usage{
grafik_geometrik_titik(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika tambahan untuk layer ini (opsional)}

\item{...}{Argumen tambahan}
}
\value{
Layer titik
}
\description{
Padanan fungsi `ggplot2::geom_point()`
}
\examples{
visualisasi_ggplot(penguin, estetika(mpg, hp)) + grafik_geometrik_titik()

}
\seealso{
[ggplot2::geom_point()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
