% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_geometrik_tren}
\alias{grafik_geometrik_tren}
\title{Menambahkan garis tren}
\usage{
grafik_geometrik_tren(mapping = NULL, method = NULL, ...)
}
\arguments{
\item{mapping}{Pemetaan estetika tambahan (opsional)}

\item{method}{Metode pemodelan seperti `"lm"`}

\item{...}{Argumen tambahan}
}
\value{
Layer garis semulus (tren)
}
\description{
Padanan fungsi `ggplot2::geom_smooth()`
}
\examples{
visualisasi_ggplot(penguin, estetika(mpg, hp)) +
  grafik_geometrik_tren(method = "lm")

}
\seealso{
[ggplot2::geom_smooth()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
