% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{grafik_label}
\alias{grafik_label}
\title{Menambahkan label dan judul pada grafik}
\usage{
grafik_label(
  title = NULL,
  subtitle = NULL,
  x = NULL,
  y = NULL,
  color = NULL,
  shape = NULL,
  ...
)
}
\arguments{
\item{title}{Judul utama}

\item{subtitle}{Subjudul}

\item{x}{Label sumbu X}

\item{y}{Label sumbu Y}

\item{color}{Label legenda warna}

\item{shape}{Label legenda bentuk}

\item{...}{Argumen tambahan}
}
\value{
Layer label
}
\description{
Padanan fungsi `ggplot2::labs()`
}
\examples{
\donttest{
visualisasi_ggplot(penguin, estetika(mpg, hp)) +
  grafik_geometrik_titik() +
  grafik_label(title = "Hubungan MPG dan HP", x = "MPG", y = "HP")
}

}
\seealso{
[ggplot2::labs()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
