% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_wrapper.R
\name{hari_minggu}
\alias{hari_minggu}
\title{Hari Minggu}
\usage{
hari_minggu(x)
}
\arguments{
\item{x}{Tanggal / datetime}
}
\value{
Hari dalam minggu (numeric)
}
\description{
Wrapper untuk lubridate::wday()
}
