% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(...)
}
\arguments{
\item{...}{Argumen tambahan untuk geom_histogram()}
}
\value{
Layer ggplot2
}
\description{
Wrapper untuk ggplot2::geom_histogram()
}
