% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impor.R
\name{impor_baca_csv}
\alias{impor_baca_csv}
\title{Mengimpor data dari file CSV}
\usage{
impor_baca_csv(path, ...)
}
\arguments{
\item{path}{Jalur atau nama file CSV}

\item{...}{Argumen tambahan untuk `readr::read_csv()`}
}
\value{
Tibble hasil pembacaan CSV
}
\description{
Padanan fungsi `readr::read_csv()`
}
\examples{
# Contoh menggunakan file CSV buatan
temp <- tempfile(fileext = ".csv")
write.csv(iris, temp, row.names = FALSE)
impor_baca_csv(temp)

}
\seealso{
[readr::read_csv()]

Other impor: 
\code{\link{impor_baca_excel}()},
\code{\link{input_tulis_csv}()},
\code{\link{input_tulis_excel}()}
}
\concept{impor}
