% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impor.R
\name{impor_baca_excel}
\alias{impor_baca_excel}
\title{Mengimpor data dari file Excel}
\usage{
impor_baca_excel(path, sheet = NULL, range = NULL, col_types = NULL, na = "NA")
}
\arguments{
\item{path}{Jalur atau nama file Excel}

\item{sheet}{Nama atau nomor sheet (default: sheet pertama)}

\item{range}{Rentang sel (opsional)}

\item{col_types}{Jenis kolom (opsional)}

\item{na}{Nilai string yang dianggap sebagai `NA`. Default: `"NA"`}
}
\value{
Sebuah data frame berisi data dari file Excel
}
\description{
Padanan fungsi `readxl::read_excel()`
}
\examples{
# Contoh minimal (file harus tersedia)
\donttest{
# impor_baca_excel("data/penjualan.xlsx")
}

}
\seealso{
[readxl::read_excel()]

Other impor: 
\code{\link{impor_baca_csv}()},
\code{\link{input_tulis_csv}()},
\code{\link{input_tulis_excel}()}
}
\concept{impor}
