% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{input_tabel_tibble}
\alias{input_tabel_tibble}
\title{Membuat tabel tibble}
\usage{
input_tabel_tibble(...)
}
\arguments{
\item{...}{Nama kolom dan nilai yang akan dimasukkan ke dalam tibble}
}
\value{
Sebuah tibble
}
\description{
Padanan fungsi `tibble::tibble()`
}
\examples{
input_tabel_tibble(
  nama = c("A", "B"),
  nilai = c(10, 20)
)

}
\seealso{
[tibble::tibble()], [input_tabel_tribble()]

Other input: 
\code{\link{input_tabel_tribble}()},
\code{\link{input_vektor}()}
}
\concept{input}
