% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{input_tabel_tribble}
\alias{input_tabel_tribble}
\title{Membuat tibble baris-per-baris (tribble)}
\usage{
input_tabel_tribble(...)
}
\arguments{
\item{...}{Nilai yang dituliskan baris per baris untuk membentuk tibble.}
}
\value{
Sebuah tibble yang disusun baris demi baris.
}
\description{
Padanan fungsi `tibble::tribble()`
}
\examples{
input_tabel_tribble(
  ~nama, ~nilai,
  "A",     10,
  "B",     20
)

}
\seealso{
[tibble::tribble()], [input_tabel_tibble()]

Other input: 
\code{\link{input_tabel_tibble}()},
\code{\link{input_vektor}()}
}
\concept{input}
