% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impor.R
\name{input_tulis_csv}
\alias{input_tulis_csv}
\title{Menyimpan data ke file CSV}
\usage{
input_tulis_csv(data, path, ...)
}
\arguments{
\item{data}{Sebuah data frame atau tibble yang akan disimpan}

\item{path}{Jalur atau nama file CSV tujuan}

\item{...}{Argumen tambahan untuk `readr::write_csv()`}
}
\value{
File CSV tersimpan ke lokasi yang ditentukan
}
\description{
Padanan fungsi `readr::write_csv()`
}
\examples{
\donttest{
# contoh data
# input_tulis_csv(mtcars, "data/mtcars.csv")
}

}
\seealso{
[readr::write_csv()]

Other impor: 
\code{\link{impor_baca_csv}()},
\code{\link{impor_baca_excel}()},
\code{\link{input_tulis_excel}()}
}
\concept{impor}
