% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impor.R
\name{input_tulis_excel}
\alias{input_tulis_excel}
\title{Menyimpan data ke file Excel}
\usage{
input_tulis_excel(data, path, ...)
}
\arguments{
\item{data}{Sebuah data frame atau tibble yang akan disimpan}

\item{path}{Jalur atau nama file Excel tujuan}

\item{...}{Argumen tambahan untuk `writexl::write_xlsx()`}
}
\value{
File Excel tersimpan ke lokasi yang ditentukan
}
\description{
Padanan fungsi `writexl::write_xlsx()`
}
\examples{
\donttest{
# contoh data
# input_tulis_excel(mtcars, "data/mtcars.xlsx")
}

}
\seealso{
[writexl::write_xlsx()]

Other impor: 
\code{\link{impor_baca_csv}()},
\code{\link{impor_baca_excel}()},
\code{\link{input_tulis_csv}()}
}
\concept{impor}
