% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasar.R
\name{input_vektor}
\alias{input_vektor}
\title{Menggabungkan nilai menjadi sebuah vektor}
\usage{
input_vektor(...)
}
\arguments{
\item{...}{Nilai-nilai yang ingin digabungkan menjadi vektor.}
}
\value{
Sebuah vektor (numeric, character, logical, atau jenis lainnya
        tergantung input).
}
\description{
Padanan fungsi `c()`
}
\examples{
input_vektor(1, 2, 3)
input_vektor("a", "b", "c")

}
\seealso{
[base::c()]

Other input: 
\code{\link{input_tabel_tibble}()},
\code{\link{input_tabel_tribble}()}
}
\concept{input}
