% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{kelompokkan}
\alias{kelompokkan}
\title{Kelompokkan}
\usage{
kelompokkan(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Kolom untuk pengelompokan}
}
\value{
Dataset yang sudah dikelompokkan
}
\description{
Wrapper untuk dplyr::group_by()
}
