% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_ambil}
\alias{kolom_ambil}
\title{Mengambil kolom menjadi vektor}
\usage{
kolom_ambil(.data, col)
}
\arguments{
\item{.data}{Data frame atau tibble yang berisi kolom yang akan diambil}

\item{col}{Kolom yang ingin ditarik. Bisa berupa nama kolom (tanpa tanda kutip)
atau posisi kolom secara numerik. Nilai negatif dapat digunakan untuk
memilih kolom berdasarkan posisi dari belakang}
}
\value{
Vektor berisi nilai dari kolom yang dipilih
}
\description{
Padanan fungsi `dplyr::pull()`
}
\examples{
library(tibble)
library(dplyr)

data <- tibble(
  nama = c("A", "B", "C"),
  nilai = c(10, 20, 30)
)

# Mengambil kolom sebagai vektor
kolom_ambil(data, nilai)

# Mengambil kolom kedua
kolom_ambil(data, 2)

# Mengambil kolom terakhir
kolom_ambil(data, -1)

}
