% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_diakhiri}
\alias{kolom_diakhiri}
\title{Memilih kolom yang namanya diakhiri dengan pola tertentu}
\usage{
kolom_diakhiri(suffix, ignore.case = TRUE)
}
\arguments{
\item{suffix}{Teks yang harus menjadi akhir nama kolom.}

\item{ignore.case}{Logika, apakah mengabaikan huruf besar/kecil.}
}
\value{
Objek tidyselect untuk memilih kolom.
}
\description{
Padanan fungsi `tidyselect::ends_with()`
}
\examples{
df <- tibble::tibble(id_user = 1:3, nama = c("a","b","c"))
transformasi_seleksi(df, kolom_diakhiri("user"))

}
