% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_diawali}
\alias{kolom_diawali}
\title{Memilih kolom yang namanya diawali dengan pola tertentu}
\usage{
kolom_diawali(prefix, ignore.case = TRUE)
}
\arguments{
\item{prefix}{Teks yang harus menjadi awal nama kolom.}

\item{ignore.case}{Logika, apakah mengabaikan huruf besar/kecil}
}
\value{
Objek tidyselect untuk memilih kolom
}
\description{
Padanan fungsi `tidyselect::starts_with()`
}
\examples{
df <- tibble::tibble(nama1 = 1:3, nama2 = 4:6, umur = 20:22)
transformasi_seleksi(df, kolom_diawali("nama"))

}
