% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{kolom_ekstrak_angka}
\alias{kolom_ekstrak_angka}
\title{Mengekstrak angka dari teks}
\usage{
kolom_ekstrak_angka(
  x,
  locale = readr::locale(),
  na = c("", "NA"),
  trim_ws = TRUE
)
}
\arguments{
\item{x}{Vektor karakter yang mengandung angka dan teks.}

\item{locale}{Lokasi numerik (opsional), mengikuti parameter `readr::parse_number()`.}

\item{na}{Pasangan teks tertentu yang dianggap sebagai `NA`.}

\item{trim_ws}{Logika: apakah spasi kosong di awal/akhir dihapus sebelum parsing.}
}
\value{
Vektor numerik hasil ekstraksi angka dari teks.
}
\description{
Padanan fungsi `readr::parse_number()`
}
\examples{
kolom_ekstrak_angka("Rp 1.500.000")
kolom_ekstrak_angka("Diskon 20\%")
kolom_ekstrak_angka(c("abc123", "45kg"))

}
\seealso{
[readr::parse_number()]

Other rapi: 
\code{\link{rapi_gabung_kolom}()},
\code{\link{rapi_hapus_na}()},
\code{\link{rapi_isi_na}()},
\code{\link{rapi_pisah_kolom}()},
\code{\link{rapi_pivot_melebar}()},
\code{\link{rapi_pivot_memanjang}()}
}
\concept{rapi}
