% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_ganti_nama}
\alias{kolom_ganti_nama}
\title{Mengganti nama kolom}
\usage{
kolom_ganti_nama(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{pasangan nama baru = nama lama}
}
\value{
Data frame dengan nama kolom baru
}
\description{
Padanan fungsi `dplyr::rename()`
}
\examples{
df <- data.frame(age = 10)
kolom_ganti_nama(df, umur = age)

}
\seealso{
[dplyr::rename()]
}
\concept{data}
