% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_mengandung}
\alias{kolom_mengandung}
\title{Memilih kolom yang namanya mengandung teks tertentu}
\usage{
kolom_mengandung(text, ignore.case = TRUE)
}
\arguments{
\item{text}{Teks yang harus muncul di dalam nama kolom.}

\item{ignore.case}{Logika, apakah mengabaikan huruf besar/kecil.}
}
\value{
Objek tidyselect.
}
\description{
Padanan fungsi `tidyselect::contains()`
}
\examples{
df <- tibble::tibble(
  nama_depan = c("A", "B", "C"),
  nama_belakang = c("X", "Y", "Z"),
  usia = c(20, 30, 40)
)

transformasi_seleksi(df, kolom_mengandung("nama"))

}
