% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_menurun}
\alias{kolom_menurun}
\title{Mengurutkan nilai secara menurun}
\usage{
kolom_menurun(x)
}
\arguments{
\item{x}{Kolom atau vektor yang akan diurutkan menurun}
}
\value{
Objek yang menandai pengurutan menurun
}
\description{
Padanan fungsi `dplyr::desc()`
}
\examples{
df <- tibble::tibble(nama = c("A", "B", "C"), nilai = c(10, 30, 20))

# Urutkan berdasarkan nilai secara menurun
transformasi_urutkan(df, kolom_menurun(nilai))

}
\seealso{
[dplyr::desc()], [transformasi_urutkan()]
}
