% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_pilih_sama}
\alias{kolom_pilih_sama}
\title{Memilih kolom berdasarkan kesamaan nama}
\usage{
kolom_pilih_sama(pattern, ignore.case = FALSE)
}
\arguments{
\item{pattern}{Pola regular expression yang akan dikolom_cocokkan.}

\item{ignore.case}{Logika, apakah mengabaikan huruf besar/kecil.}
}
\value{
Objek tidyselect.
}
\description{
Padanan fungsi `tidyselect::matches()`.
}
\examples{
df <- tibble::tibble(
  x1 = 1:3,
  x2 = 3:5,
  y = 10:12
)

transformasi_seleksi(df, kolom_cocok("^x[0-9]+"))

}
