% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_pilih_semua}
\alias{kolom_pilih_semua}
\title{Memilih semua kolom}
\usage{
kolom_pilih_semua()
}
\value{
Objek tidyselect yang mewakili semua kolom
}
\description{
Padanan fungsi `tidyselect::everything()`
}
\examples{
df <- tibble::tibble(a = 1, b = 2, c = 3)
transformasi_seleksi(df, kolom_pilih_semua())

}
