% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_rentang_angka}
\alias{kolom_rentang_angka}
\title{Memilih kolom berdasarkan pola rentang angka}
\usage{
kolom_rentang_angka(prefix, range, width = NULL)
}
\arguments{
\item{prefix}{Nama awal kolom}

\item{range}{Vektor angka, misalnya 1:5}

\item{width}{Jumlah digit, misalnya 2 untuk x01, x02}
}
\value{
Objek tidyselect.
}
\description{
Padanan fungsi `tidyselect::num_range()`
}
\examples{
df <- tibble::tibble(x1 = 1, x2 = 2, x3 = 3)
transformasi_seleksi(df, kolom_rentang_angka("x", 1:3))

}
