% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantuan.R
\name{kolom_tipe}
\alias{kolom_tipe}
\title{Memilih kolom berdasarkan kondisi fungsi}
\usage{
kolom_tipe(fn)
}
\arguments{
\item{fn}{Fungsi predicate yang mengembalikan TRUE/FALSE}
}
\value{
Objek tidyselect.
}
\description{
Padanan fungsi `tidyselect::where()`
}
\details{
Fungsi ini hanya digunakan dalam konteks seleksi seperti
`transformasi_seleksi()`
}
\examples{
df <- tibble::tibble(a = 1:3, b = 4:6, c = c("x", "y", "z"))
transformasi_seleksi(df, kolom_tipe(is.numeric))

}
