% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat}
\alias{lihat}
\title{Menampilkan data frame dalam tampilan spreadsheet RStudio atau GUI}
\usage{
lihat(data)
}
\arguments{
\item{data}{Data frame yang ingin dilihat.}
}
\value{
Membuka tampilan data di panel RStudio atau jendela baru.
}
\description{
Fungsi ini menampilkan data frame dalam tampilan interaktif seperti lembar kerja.
Jika dijalankan di RStudio, data akan muncul di panel Data Viewer dengan nama
objek asli sebagai judul tab. Jika dijalankan di luar RStudio, data akan muncul
di jendela baru (GUI).
}
\examples{
# lihat(penguins)  # menampilkan data di panel Data Viewer (RStudio)
}
\seealso{
[utils::View()]
}
