% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat_dimensi}
\alias{lihat_dimensi}
\title{Melihat jumlah baris dan kolom dari data frame.}
\usage{
lihat_dimensi(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Vektor numerik berisi jumlah baris dan kolom.
}
\description{
Fungsi ini adalah padanan dari `dim()` dan digunakan untuk mengetahui
ukuran data (baris x kolom).
}
\examples{
lihat_dimensi(penguins)
}
\seealso{
[base::dim()]
}
