% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat_direktori_kerja}
\alias{lihat_direktori_kerja}
\title{Menampilkan direktori kerja saat ini}
\usage{
lihat_direktori_kerja()
}
\value{
String yang berisi path direktori kerja saat ini.
}
\description{
Fungsi ini menampilkan direktori kerja aktif di R.
Ini adalah padanan dari `getwd()` dan berguna untuk mengetahui
lokasi folder saat ini di mana R membaca dan menyimpan file.
}
\examples{
# lihat_direktori_kerja()
# [1] "C:/Users/Username/Documents/R"
}
\seealso{
[getwd()]
}
