% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat_ekor}
\alias{lihat_ekor}
\title{Menampilkan beberapa baris terakhir dari data frame.}
\usage{
lihat_ekor(data, n = 6)
}
\arguments{
\item{data}{Data frame atau vektor.}

\item{n}{Jumlah baris yang ingin ditampilkan (default = 6).}
}
\value{
Menampilkan beberapa baris terakhir dari data.
}
\description{
Fungsi ini adalah padanan dari `tail()` dan digunakan untuk melihat
bagian akhir dari data.
}
\examples{
lihat_ekor(penguins, 3)
}
\seealso{
[base::tail()]
}
