% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat_kepala}
\alias{lihat_kepala}
\title{Menampilkan beberapa baris pertama dari data frame.}
\usage{
lihat_kepala(data, n = 6)
}
\arguments{
\item{data}{Data frame atau vektor.}

\item{n}{Jumlah baris yang ingin ditampilkan (default = 6).}
}
\value{
Menampilkan beberapa baris pertama dari data.
}
\description{
Fungsi ini adalah padanan dari `head()` dan digunakan untuk melihat
contoh awal dari data.
}
\examples{
lihat_kepala(penguins, 3)
}
\seealso{
[base::head()]
}
