% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat_nama_kolom}
\alias{lihat_nama_kolom}
\title{Melihat nama kolom dari data frame.}
\usage{
lihat_nama_kolom(data)

lihat_nama_kolom(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Karakter vektor berisi nama kolom.

Karakter vektor berisi nama kolom.
}
\description{
Fungsi ini adalah padanan dari `names()` dan digunakan untuk menampilkan
daftar nama kolom yang terdapat pada data.

Fungsi ini adalah padanan dari `names()` dan digunakan untuk menampilkan
daftar nama kolom yang terdapat pada data.
}
\examples{
lihat_nama_kolom(penguins)
lihat_nama_kolom(penguins)
}
\seealso{
[base::names()]

[base::names()]
}
