% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{lihat_sekilas}
\alias{lihat_sekilas}
\title{Melihat sekilas struktur data frame dengan cara ringkas dan rapi.}
\usage{
lihat_sekilas(data)
}
\arguments{
\item{data}{Data frame yang ingin dilihat strukturnya.}
}
\value{
Menampilkan ringkasan struktur data di konsol.
}
\description{
Fungsi ini adalah padanan dari `dplyr::glimpse()` dan digunakan untuk
menampilkan struktur data secara cepat tanpa mencetak seluruh isi data.
}
\examples{
lihat_sekilas(penguins)
}
\seealso{
[dplyr::glimpse()]
}
