% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mengandung}
\alias{mengandung}
\title{Mengandung}
\usage{
mengandung(match)
}
\arguments{
\item{match}{String yang dicari di nama kolom}
}
\value{
Kolom yang sesuai
}
\description{
Wrapper untuk dplyr::contains()
}
