% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{model_regresi}
\alias{model_regresi}
\title{Membuat Model Regresi Linear}
\usage{
model_regresi(formula, data, ...)
}
\arguments{
\item{formula}{Rumus model, misalnya `y ~ x1 + x2`.}

\item{data}{Data frame yang digunakan untuk pemodelan.}

\item{...}{Argumen tambahan untuk `stats::lm()`.}
}
\value{
Objek model regresi linear.
}
\description{
Membuat model regresi linear untuk menganalisis hubungan antara variabel
respon dan satu atau lebih variabel prediktor.
Fungsi ini adalah padanan dari `stats::lm()`.
}
\examples{
# model <- model_regresi(body_mass_g ~ flipper_length_mm, data = penguins)
# ringkasan_model(model)

}
\seealso{
[stats::lm()]
}
