% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{panel_bungkus}
\alias{panel_bungkus}
\title{Membagi plot menjadi beberapa panel berdasarkan variabel}
\usage{
panel_bungkus(facets, ...)
}
\arguments{
\item{facets}{Variabel untuk pemisahan subplot.}

\item{...}{Argumen tambahan untuk `ggplot2::facet_wrap()` atau `ggplot2::facet_grid()`.}
}
\description{
Membagi visualisasi ke dalam beberapa subplot berdasarkan satu atau dua variabel kategorik.
Fungsi ini adalah padanan dari `ggplot2::facet_wrap()` dan `ggplot2::facet_grid()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = flipper_length_mm, y = body_mass_g)) +
#   geometrik_titik() +
#   panel_bungkus(~ species)

}
\seealso{
[ggplot2::facet_wrap()], [ggplot2::facet_grid()]
}
