% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{panel_grid}
\alias{panel_grid}
\title{Membagi plot menjadi grid berdasarkan dua variabel}
\usage{
panel_grid(rows = NULL, cols = NULL, ...)
}
\arguments{
\item{rows, cols}{Variabel untuk baris dan kolom.}

\item{...}{Argumen tambahan untuk `ggplot2::facet_grid()`.}
}
\description{
Padanan dari `ggplot2::facet_grid()`.
}
\examples{
# visualisasi_ggplot(penguins, estetika(x = flipper_length_mm, y = body_mass_g)) +
#   geometrik_titik() +
#   panel_grid(rows = vars(species), cols = vars(sex))

}
\seealso{
[ggplot2::facet_grid()]
}
