% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{rapi_gabung_kolom}
\alias{rapi_gabung_kolom}
\title{Menggabungkan beberapa kolom menjadi satu}
\usage{
rapi_gabung_kolom(data, col, ..., sep = "_")
}
\arguments{
\item{data}{Data frame.}

\item{col}{Kolom baru.}

\item{...}{Kolom yang digabungkan.}

\item{sep}{Pemisah.}
}
\value{
Data frame dengan kolom baru hasil penggabungan.
}
\description{
Padanan fungsi `tidyr::unite()`.
}
\examples{
df <- data.frame(tahun = 2020, bulan = "01")
rapi_gabung_kolom(df, tanggal, tahun, bulan, sep = "-")

}
\seealso{
[tidyr::unite()]

Other rapi: 
\code{\link{kolom_ekstrak_angka}()},
\code{\link{rapi_hapus_na}()},
\code{\link{rapi_isi_na}()},
\code{\link{rapi_pisah_kolom}()},
\code{\link{rapi_pivot_melebar}()},
\code{\link{rapi_pivot_memanjang}()}
}
\concept{rapi}
