% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_ganti_nama}
\alias{rapi_ganti_nama}
\title{Mengganti nama kolom dalam data}
\usage{
rapi_ganti_nama(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Pasangan nama baru = nama lama.}
}
\value{
Data frame dengan nama kolom yang diperbarui.
}
\description{
Mengganti nama satu atau beberapa kolom. Fungsi ini adalah padanan dari `dplyr::rename()`.
}
\examples{
# rapi_ganti_nama(data, umur_baru = age)

}
\seealso{
[dplyr::rename()]
}
