% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_hapus_duplikat}
\alias{rapi_hapus_duplikat}
\title{Menghapus baris duplikat dalam data}
\usage{
rapi_hapus_duplikat(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kolom yang digunakan untuk mendeteksi duplikat (opsional).}
}
\value{
Data frame tanpa baris duplikat.
}
\description{
Menghapus baris yang memiliki nilai identik pada kolom tertentu.
Fungsi ini adalah padanan dari `dplyr::distinct()`.
}
\examples{
# rapi_hapus_duplikat(data)

}
\seealso{
[dplyr::distinct()]
}
