% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{rapi_hapus_na}
\alias{rapi_hapus_na}
\title{Menghapus baris dengan nilai hilang (NA)}
\usage{
rapi_hapus_na(data, ...)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Kolom opsional untuk diperiksa.}
}
\value{
Data frame tanpa baris NA.
}
\description{
Padanan fungsi `tidyr::drop_na()`
}
\examples{
df <- data.frame(a = c(1, NA, 3))
rapi_hapus_na(df)

}
\seealso{
[tidyr::drop_na()]

Other rapi: 
\code{\link{kolom_ekstrak_angka}()},
\code{\link{rapi_gabung_kolom}()},
\code{\link{rapi_isi_na}()},
\code{\link{rapi_pisah_kolom}()},
\code{\link{rapi_pivot_melebar}()},
\code{\link{rapi_pivot_memanjang}()}
}
\concept{rapi}
