% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{rapi_isi_na}
\alias{rapi_isi_na}
\title{Mengisi nilai hilang (NA)}
\usage{
rapi_isi_na(data, ..., .direction = "down")
}
\arguments{
\item{data}{Data frame.}

\item{...}{Kolom untuk diisi.}

\item{.direction}{Arah pengisian.}
}
\value{
Data frame dengan NA yang sudah diisi.
}
\description{
Padanan fungsi `tidyr::fill()`
}
\examples{
df <- data.frame(a = c(1, NA, 3))
rapi_isi_na(df, a, .direction = "down")

}
\seealso{
[tidyr::fill()]

Other rapi: 
\code{\link{kolom_ekstrak_angka}()},
\code{\link{rapi_gabung_kolom}()},
\code{\link{rapi_hapus_na}()},
\code{\link{rapi_pisah_kolom}()},
\code{\link{rapi_pivot_melebar}()},
\code{\link{rapi_pivot_memanjang}()}
}
\concept{rapi}
