% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{rapi_pisah_kolom}
\alias{rapi_pisah_kolom}
\title{Memisahkan satu kolom menjadi beberapa kolom}
\usage{
rapi_pisah_kolom(data, col, into, sep, ...)
}
\arguments{
\item{data}{Data frame.}

\item{col}{Kolom yang ingin dipisah.}

\item{into}{Nama kolom hasil.}

\item{sep}{Pola pemisah.}

\item{...}{Argumen tambahan.}
}
\value{
Data frame dengan kolom baru hasil pemisahan.
}
\description{
Padanan fungsi `tidyr::separate()`.
}
\examples{
df <- data.frame(tanggal = c("2020-01", "2020-02"))
rapi_pisah_kolom(df, tanggal, into = c("tahun", "bulan"), sep = "-")

}
\seealso{
[tidyr::separate()]

Other rapi: 
\code{\link{kolom_ekstrak_angka}()},
\code{\link{rapi_gabung_kolom}()},
\code{\link{rapi_hapus_na}()},
\code{\link{rapi_isi_na}()},
\code{\link{rapi_pivot_melebar}()},
\code{\link{rapi_pivot_memanjang}()}
}
\concept{rapi}
