% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{rapi_pivot_melebar}
\alias{rapi_pivot_melebar}
\title{Merapikan data menjadi format melebar (menambah kolom)}
\usage{
rapi_pivot_melebar(data, names_from, values_from, ...)
}
\arguments{
\item{data}{Data frame}

\item{names_from}{Kolom untuk nama variabel}

\item{values_from}{Kolom untuk nilai}

\item{...}{Argumen tambahan}
}
\value{
Data frame dalam format lebar
}
\description{
Padanan fungsi `tidyr::pivot_wider()`
}
\examples{
df <- data.frame(bulan = c("jan","feb"), nilai = c(10, 20))
rapi_pivot_melebar(df, names_from = bulan, values_from = nilai)

}
\seealso{
Other rapi: 
\code{\link{kolom_ekstrak_angka}()},
\code{\link{rapi_gabung_kolom}()},
\code{\link{rapi_hapus_na}()},
\code{\link{rapi_isi_na}()},
\code{\link{rapi_pisah_kolom}()},
\code{\link{rapi_pivot_memanjang}()}
}
\concept{rapi}
