% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapi.R
\name{rapi_pivot_memanjang}
\alias{rapi_pivot_memanjang}
\title{Merapikan data menjadi format memanjang (menambah baris)}
\usage{
rapi_pivot_memanjang(data, cols, names_to = "name", values_to = "value", ...)
}
\arguments{
\item{data}{Data frame.}

\item{cols}{Kolom yang diputar.}

\item{names_to}{Nama kolom baru.}

\item{values_to}{Nama kolom nilai.}

\item{...}{Argumen tambahan.}
}
\value{
Data frame dalam format memanjang.
}
\description{
Padanan fungsi `tidyr::pivot_longer()`
}
\examples{
df <- data.frame(a = 1:3, jan = 1:3, feb = 4:6)
rapi_pivot_memanjang(df, cols = c(jan, feb), names_to = "bulan", values_to = "nilai")

}
\seealso{
[tidyr::pivot_longer()]

Other rapi: 
\code{\link{kolom_ekstrak_angka}()},
\code{\link{rapi_gabung_kolom}()},
\code{\link{rapi_hapus_na}()},
\code{\link{rapi_isi_na}()},
\code{\link{rapi_pisah_kolom}()},
\code{\link{rapi_pivot_melebar}()}
}
\concept{rapi}
