% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_urut_faktor}
\alias{rapi_urut_faktor}
\title{Mengurutkan faktor berdasarkan frekuensi kemunculan}
\usage{
rapi_urut_faktor(f, ...)
}
\arguments{
\item{f}{Vektor faktor.}

\item{...}{Argumen tambahan untuk `forcats::fct_infreq()`.}
}
\value{
Faktor dengan urutan level berdasarkan frekuensi.
}
\description{
Mengurutkan level faktor dari yang paling sering muncul hingga paling jarang.
Fungsi ini adalah padanan dari `forcats::fct_infreq()`.
}
\examples{
# rapi_urut_faktor(species)

}
\seealso{
[forcats::fct_infreq()]
}
