% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ringkas}
\alias{ringkas}
\title{Ringkas}
\usage{
ringkas(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Fungsi ringkasan}
}
\value{
Dataset yang diringkas
}
\description{
Wrapper untuk dplyr::summarise()
}
