% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{tema_gelap}
\alias{tema_gelap}
\title{Tema Gelap}
\usage{
tema_gelap()
}
\value{
Layer ggplot untuk digunakan di +.
}
\description{
Padanan dari `ggplot2::theme_dark()` untuk plot dengan latar gelap.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) +
  ggplot2::geom_point() +
  tema_gelap()
}
\seealso{
[ggplot2::theme_dark()], [tema()]
}
