% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{tema_klasik}
\alias{tema_klasik}
\title{Tema Klasik}
\usage{
tema_klasik()
}
\value{
Layer ggplot untuk digunakan di +.
}
\description{
Padanan dari `ggplot2::theme_classic()` untuk tampilan plot klasik dengan grid minimal.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) +
  ggplot2::geom_point() +
  tema_klasik()
}
\seealso{
[ggplot2::theme_classic()], [tema()]
}
