% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{tema_kosong}
\alias{tema_kosong}
\title{Tema Kosong}
\usage{
tema_kosong()
}
\value{
Layer ggplot untuk digunakan di +.
}
\description{
Padanan dari `ggplot2::theme_void()` untuk plot tanpa latar dan sumbu.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) +
  ggplot2::geom_point() +
  tema_kosong()
}
\seealso{
[ggplot2::theme_void()], [tema()]
}
