% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{tema_minimal}
\alias{tema_minimal}
\title{Tema Minimal}
\usage{
tema_minimal()
}
\value{
Layer ggplot untuk digunakan di +.
}
\description{
Padanan dari `ggplot2::theme_minimal()` untuk tampilan plot yang bersih.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) +
  ggplot2::geom_point() +
  tema_minimal()
}
\seealso{
[ggplot2::theme_minimal()], [tema()]
}
