% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{tema_sederhana}
\alias{tema_sederhana}
\title{Tema Sederhana}
\usage{
tema_sederhana()
}
\value{
Layer ggplot untuk digunakan di +.
}
\description{
Padanan dari `ggplot2::theme_light()` untuk plot dengan grid lembut dan latar putih.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) +
  ggplot2::geom_point() +
  tema_sederhana()
}
\seealso{
[ggplot2::theme_light()], [tema()]
}
