% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_wrapper.R
\name{thnblntgl}
\alias{thnblntgl}
\title{thnblntgl}
\usage{
thnblntgl(x)
}
\arguments{
\item{x}{Karakter yang berisi tanggal dalam format tahun-bulan-tanggal, misal "2025-10-27"}
}
\value{
Objek Date
}
\description{
Mengubah tanggal dari format "YYYY-MM-DD" menjadi objek Date.
}
