% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paket.R
\docType{package}
\name{tidyversindonesia}
\alias{tidyversindonesia-package}
\alias{tidyversindonesia}
\title{Paket tidyverse versi Bahasa Indonesia}
\description{
Menyediakan padanan fungsi inti tidyverse dalam Bahasa Indonesia
sebagai jembatan untuk mempelajari sains data menggunakan R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yandiandiyana/tidyversindonesia}
}

}
\author{
\strong{Maintainer}: Yandi Andiyana \email{tidyversindonesia@gmail.com}

}
