% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_distingsi}
\alias{transformasi_distingsi}
\title{Memilih hanya baris berbeda (distingsi) tanpa duplikat}
\usage{
transformasi_distingsi(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{Kolom opsional}
}
\value{
Data frame berbeda (distingsi) tanpa duplikat
}
\description{
Padanan fungsi `dplyr::distinct()`
}
\examples{
df <- data.frame(a = c(1,1,2))
transformasi_distingsi(df)

}
\seealso{
[dplyr::distinct()]

Other transformasi: 
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
