% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_faktor_menurun}
\alias{transformasi_faktor_menurun}
\title{Mengurutkan faktor berdasarkan frekuensi kemunculan}
\usage{
transformasi_faktor_menurun(f, ...)
}
\arguments{
\item{f}{Vektor faktor}

\item{...}{Argumen tambahan}
}
\value{
Faktor dengan urutan level berdasarkan frekuensi.
}
\description{
Padanan fungsi `forcats::fct_infreq()`
}
\examples{
f <- factor(c("A","B","A","C"))
transformasi_faktor_menurun(f)

}
\seealso{
[forcats::fct_infreq()]
}
