% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_filter}
\alias{transformasi_filter}
\title{Menyaring baris berdasarkan kondisi tertentu}
\usage{
transformasi_filter(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{...}{Kondisi logika untuk menyaring baris}
}
\value{
Data frame hasil penyaringan
}
\description{
Padanan fungsi `dplyr::filter()`
}
\examples{
df <- tibble::tibble(nama = c("A", "B"), usia = c(20, 30))
transformasi_filter(df, usia > 25)

}
\seealso{
[dplyr::filter()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
