% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_kelompok_berdasarkan}
\alias{transformasi_kelompok_berdasarkan}
\title{Mengelompokkan data berdasarkan kolom tertentu}
\usage{
transformasi_kelompok_berdasarkan(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{...}{Kolom yang digunakan untuk pengelompokan}
}
\value{
Data frame bertipe `grouped_df`
}
\description{
Padanan fungsi `dplyr::group_by()`
}
\examples{
df <- tibble::tibble(jk = c("L", "P", "L"), usia = c(20, 22, 25))
transformasi_kelompok_berdasarkan(df, jk)

}
\seealso{
[dplyr::group_by()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
