% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_kelompok_hapus}
\alias{transformasi_kelompok_hapus}
\title{Melepas pengelompokan dari data}
\usage{
transformasi_kelompok_hapus(data)
}
\arguments{
\item{data}{Data frame atau tibble yang sudah dikelompokkan}
}
\value{
Data frame tanpa pengelompokan
}
\description{
Padanan fungsi `dplyr::ungroup()`
}
\examples{
df <- tibble::tibble(jk = c("L", "P", "L"), usia = c(20, 22, 25))
g <- transformasi_kelompok_berdasarkan(df, jk)
transformasi_kelompok_hapus(g)

}
\seealso{
[dplyr::ungroup()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
