% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_keluar_grup}
\alias{transformasi_keluar_grup}
\title{Melepas pengelompokan dari data}
\usage{
transformasi_keluar_grup(data)
}
\arguments{
\item{data}{Data frame atau tibble yang sudah dikelompokkan.}
}
\value{
Data frame tanpa pengelompokan.
}
\description{
Menghapus struktur pengelompokan agar operasi berikutnya tidak per grup.
Fungsi ini adalah padanan dari `dplyr::ungroup()`.
}
\examples{
# transformasi_keluar_grup(data)

}
\seealso{
[dplyr::ungroup()]
}
