% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_masuk_grup}
\alias{transformasi_masuk_grup}
\title{Mengelompokkan data berdasarkan kolom tertentu}
\usage{
transformasi_masuk_grup(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kolom yang digunakan untuk pengelompokan.}
}
\value{
Data frame yang telah dikelompokkan.
}
\description{
Mengelompokkan data agar perhitungan atau ringkasan dilakukan per kelompok.
Fungsi ini adalah padanan dari `dplyr::group_by()`.
}
\examples{
# transformasi_masuk_grup(data, jenis_kelamin)

}
\seealso{
[dplyr::group_by()]
}
