% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_mutasi}
\alias{transformasi_mutasi}
\title{Membuat atau mengubah kolom dalam data}
\usage{
transformasi_mutasi(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{...}{Ekspresi untuk membuat atau mengubah kolom}
}
\value{
Data frame dengan kolom baru atau yang telah diperbarui
}
\description{
Padanan fungsi `dplyr::mutate()`
}
\examples{
df <- tibble::tibble(usia_bulan = c(120, 180))
transformasi_mutasi(df, usia_tahun = usia_bulan / 12)

}
\seealso{
[dplyr::mutate()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
