% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_pilih}
\alias{transformasi_pilih}
\title{Memilih kolom tertentu dari data}
\usage{
transformasi_pilih(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kolom yang ingin dipilih.}
}
\value{
Data frame dengan kolom yang dipilih.
}
\description{
Memilih kolom yang relevan dari data. Fungsi ini adalah padanan dari `dplyr::select()`.
}
\examples{
# transformasi_pilih(data, nama, usia)

}
\seealso{
[dplyr::select()]
}
