% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_potong_acak}
\alias{transformasi_potong_acak}
\title{Memotong baris secara acak}
\usage{
transformasi_potong_acak(
  data,
  n = NULL,
  prop = NULL,
  weight_by = NULL,
  replace = FALSE
)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{n}{Jumlah baris acak}

\item{prop}{Proporsi baris acak}

\item{weight_by}{Kolom pembobot}

\item{replace}{Sampling dengan pengulangan atau tidak}
}
\value{
Data frame
}
\description{
Padanan fungsi `dplyr::slice_sample()`
}
\seealso{
Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
