% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_potong_ekor}
\alias{transformasi_potong_ekor}
\title{Memotong baris bagian akhir (ekor)}
\usage{
transformasi_potong_ekor(data, n = NULL, prop = NULL)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{n}{Jumlah baris}

\item{prop}{Proporsi baris}
}
\value{
Data frame
}
\description{
Padanan fungsi `dplyr::slice_tail()`
}
\seealso{
Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
