% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_potong_maksimum}
\alias{transformasi_potong_maksimum}
\title{Memotong baris berdasarkan nilai maksimum kolom}
\usage{
transformasi_potong_maksimum(
  data,
  order_by,
  n = NULL,
  prop = NULL,
  with_ties = TRUE
)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{order_by}{Kolom acuan}

\item{n}{Jumlah baris}

\item{prop}{Proporsi baris}

\item{with_ties}{Sertakan nilai yang sama.}
}
\value{
Data frame
}
\description{
Padanan fungsi `dplyr::slice_max()`
}
\seealso{
Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
