% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_relokasi}
\alias{transformasi_relokasi}
\title{Relokasi kolom dalam data}
\usage{
transformasi_relokasi(data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{data}{Data frame}

\item{...}{Kolom yang ingin dipindahkan}

\item{.before}{Kolom tujuan sebelum posisi relokasi}

\item{.after}{Kolom tujuan setelah posisi relokasi}
}
\value{
Data frame dengan kolom yang sudah dipindah
}
\description{
Padanan fungsi `dplyr::relocate()`
}
\seealso{
Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
