% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_seleksi}
\alias{transformasi_seleksi}
\title{Memilih kolom tertentu dari data}
\usage{
transformasi_seleksi(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{...}{Kolom yang ingin dipilih}
}
\value{
Data frame dengan kolom yang dipilih
}
\description{
Padanan fungsi `dplyr::select()`
}
\examples{
df <- tibble::tibble(nama = c("A", "B"), usia = c(20, 25))
transformasi_seleksi(df, nama)

}
\seealso{
[dplyr::select()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_statistik_ringkas}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
