% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_statistik_ringkas}
\alias{transformasi_statistik_ringkas}
\title{Merangkum data menjadi nilai statistik ringkas}
\usage{
transformasi_statistik_ringkas(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble (biasanya dikelompokkan)}

\item{...}{Ekspresi ringkasan yang ingin dihitung}
}
\value{
Data frame hasil peringkasan
}
\description{
Padanan fungsi `dplyr::summarize()`
}
\examples{
df <- tibble::tibble(usia = c(20, 30, 40))
transformasi_statistik_ringkas(df, rata = mean(usia))

}
\seealso{
[dplyr::summarize()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_urutkan}()}
}
\concept{transformasi}
