% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_urutkan}
\alias{transformasi_urutkan}
\title{Mengurutkan baris dalam data}
\usage{
transformasi_urutkan(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble}

\item{...}{Kolom yang digunakan untuk pengurutan}
}
\value{
Data frame yang telah diurutkan
}
\description{
Padanan fungsi `dplyr::arrange()`
}
\examples{
df <- tibble::tibble(nama = c("A", "B"), usia = c(25, 20))
transformasi_urutkan(df, usia)

}
\seealso{
[dplyr::arrange()]

Other transformasi: 
\code{\link{transformasi_distingsi}()},
\code{\link{transformasi_filter}()},
\code{\link{transformasi_kelompok_berdasarkan}()},
\code{\link{transformasi_kelompok_hapus}()},
\code{\link{transformasi_mutasi}()},
\code{\link{transformasi_potong}()},
\code{\link{transformasi_potong_acak}()},
\code{\link{transformasi_potong_ekor}()},
\code{\link{transformasi_potong_kepala}()},
\code{\link{transformasi_potong_maksimum}()},
\code{\link{transformasi_potong_minimum}()},
\code{\link{transformasi_relokasi}()},
\code{\link{transformasi_seleksi}()},
\code{\link{transformasi_statistik_ringkas}()}
}
\concept{transformasi}
