% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ubah}
\alias{ubah}
\title{Ubah}
\usage{
ubah(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Kolom baru / transformasi}
}
\value{
Dataset dengan kolom baru / diubah
}
\description{
Wrapper untuk dplyr::mutate()
}
