% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{ubah_direktori_kerja}
\alias{ubah_direktori_kerja}
\title{Mengubah direktori kerja aktif}
\usage{
ubah_direktori_kerja(path)
}
\arguments{
\item{path}{String yang berisi path direktori baru.}
}
\value{
Path direktori kerja baru setelah diubah.
}
\description{
Fungsi ini mengubah direktori kerja aktif di R.
Ini adalah padanan dari `setwd()` dan berguna untuk menentukan
lokasi folder di mana R akan membaca dan menyimpan file.
}
\examples{
\dontrun{
# ubah_direktori_kerja("C:/Users/Username/Documents/Proyek")
lihat_direktori_kerja()  # memastikan direktori sudah berubah
}
}
\seealso{
[setwd()], [getwd()], [lihat_direktori_kerja()]
}
