% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{urutkan}
\alias{urutkan}
\title{Urutkan}
\usage{
urutkan(.data, ...)
}
\arguments{
\item{.data}{Dataset}

\item{...}{Kolom untuk diurutkan}
}
\value{
Dataset diurutkan
}
\description{
Wrapper untuk dplyr::arrange()
}
