% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{visualisasi_ggplot}
\alias{visualisasi_ggplot}
\title{Membuat grafik/plot dari ggplot}
\usage{
visualisasi_ggplot(data, mapping = NULL, ...)
}
\arguments{
\item{data}{Data frame atau tibble yang akan divisualisasikan.}

\item{mapping}{Pemetaan estetika, biasanya dibuat dengan `estetika()`.}

\item{...}{Argumen tambahan untuk `ggplot2::ggplot()`.}
}
\value{
Objek `ggplot` yang dapat ditambahkan layer visualisasi.
}
\description{
Padanan fungsi `ggplot2::ggplot()`
}
\examples{
# Grafik kosong
visualisasi_ggplot(penguin)

# Grafik dengan pemetaan estetika
visualisasi_ggplot(penguin, estetika(x = mpg, y = hp))

}
\seealso{
[ggplot2::ggplot()]

Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot_simpan}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
