% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{visualisasi_ggplot_simpan}
\alias{visualisasi_ggplot_simpan}
\title{Menyimpan visualisasi ke file}
\usage{
visualisasi_ggplot_simpan(
  filename,
  plot = ggplot2::last_plot(),
  width = NULL,
  height = NULL,
  dpi = 300,
  ...
)
}
\arguments{
\item{filename}{Nama file}

\item{plot}{Objek ggplot}

\item{width, height}{Ukuran gambar (inci).}

\item{dpi}{Resolusi}

\item{...}{Argumen tambahan}
}
\value{
File gambar tersimpan
}
\description{
Padanan fungsi `ggplot2::ggsave()`
}
\examples{
\dontrun{
p <- visualisasi_ggplot(penguin, estetika(mpg, hp)) + grafik_geometrik_titik()
visualisasi_ggplot_simpan("contoh_plot.png", plot = p, width = 4, height = 3)
}

}
\seealso{
Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_terakhir}()}
}
\concept{visualisasi}
