% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{visualisasi_ggplot_terakhir}
\alias{visualisasi_ggplot_terakhir}
\title{Menampilkan plot ggplot terakhir}
\usage{
visualisasi_ggplot_terakhir()
}
\value{
Objek ggplot terakhir
}
\description{
Padanan fungsi `ggplot2::last_plot()`
}
\examples{
\donttest{
visualisasi_ggplot(penguin, estetika(mpg, hp)) + grafik_geometrik_titik()
visualisasi_ggplot_terakhir()
}

}
\seealso{
Other visualisasi: 
\code{\link{estetika}()},
\code{\link{grafik_facet_baris_kolom}()},
\code{\link{grafik_facet_kolom}()},
\code{\link{grafik_geometrik_batang}()},
\code{\link{grafik_geometrik_boxplot}()},
\code{\link{grafik_geometrik_densitas}()},
\code{\link{grafik_geometrik_histogram}()},
\code{\link{grafik_geometrik_titik}()},
\code{\link{grafik_geometrik_tren}()},
\code{\link{grafik_label}()},
\code{\link{grafik_tema_gelap}()},
\code{\link{grafik_tema_hampa}()},
\code{\link{grafik_tema_klasik}()},
\code{\link{grafik_tema_minimal}()},
\code{\link{grafik_tema_ringan}()},
\code{\link{visualisasi_ggplot}()},
\code{\link{visualisasi_ggplot_simpan}()}
}
\concept{visualisasi}
