% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{estetika}
\alias{estetika}
\title{Menentukan pemetaan estetika visualisasi}
\usage{
estetika(...)
}
\arguments{
\item{...}{Pemetaan estetika seperti `x`, `y`, `color`, `shape`, dan `size`.}
}
\value{
Objek pemetaan estetika (`aes`).
}
\description{
Menentukan kolom data yang digunakan sebagai sumbu x, y, warna, bentuk, dan ukuran.
Fungsi ini adalah padanan dari `ggplot2::aes()`.
}
\examples{
# estetika(x = flipper_length_mm, y = body_mass_g)

}
\seealso{
[ggplot2::aes()]
}
