% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{impor_baca_csv}
\alias{impor_baca_csv}
\title{Impor data dari file CSV}
\usage{
impor_baca_csv(path, ...)
}
\arguments{
\item{path}{Jalur atau nama file CSV yang akan dibaca.}

\item{...}{Argumen tambahan yang diteruskan ke `readr::read_csv()`.}
}
\value{
Data frame (`tibble`) hasil pembacaan file.
}
\description{
Mengimpor dan membaca file CSV ke dalam R dengan sintaks Tidyverse versi Bahasa Indonesia.
Fungsi ini adalah padanan dari `readr::read_csv()`.
}
\examples{
# impor_baca_csv("data/contoh.csv")

}
\seealso{
[readr::read_csv()]
}
