% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{impor_baca_excel}
\alias{impor_baca_excel}
\title{Impor Data dari File Excel}
\usage{
impor_baca_excel(path, sheet = NULL, range = NULL, col_types = NULL, na = "NA")
}
\arguments{
\item{path}{Jalur atau nama file Excel yang akan dibaca.}

\item{sheet}{Nama atau nomor lembar kerja yang akan diimpor. Default-nya adalah lembar pertama.}

\item{range}{Rentang sel (opsional) yang akan dibaca, misalnya `"A1:D10"`.}

\item{col_types}{Jenis kolom yang akan dibaca (opsional).}

\item{na}{Nilai string yang akan dianggap sebagai `NA`. Default-nya adalah `"NA"`.}
}
\value{
Sebuah data frame yang berisi data dari file Excel.
}
\description{
Mengimpor dan membaca file Excel (.xls atau .xlsx) serta mengubahnya menjadi data frame.
Fungsi ini adalah padanan dari `readxl::read_excel()`.
}
\examples{
# impor_baca_excel("data/penjualan.xlsx", sheet = 1)

}
\seealso{
[readxl::read_excel()]
}
