% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_gabung_kolom}
\alias{rapi_gabung_kolom}
\title{Menggabungkan beberapa kolom menjadi satu}
\usage{
rapi_gabung_kolom(data, col, ..., sep = "_")
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{col}{Nama kolom baru hasil penggabungan.}

\item{...}{Kolom yang akan digabungkan.}

\item{sep}{Pemisah antar nilai kolom (default: "_").}
}
\value{
Data frame dengan kolom baru hasil penggabungan.
}
\description{
Menggabungkan dua atau lebih kolom menjadi satu kolom baru.
Fungsi ini adalah padanan dari `tidyr::unite()`.
}
\examples{
# rapi_gabung_kolom(data, tanggal, tahun, bulan, sep = "-")

}
\seealso{
[tidyr::unite()]
}
