% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_hapus_na}
\alias{rapi_hapus_na}
\title{Menghapus baris dengan nilai hilang (NA)}
\usage{
rapi_hapus_na(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kolom yang akan diperiksa (opsional).}
}
\value{
Data frame tanpa baris yang mengandung NA.
}
\description{
Menghapus seluruh baris yang memiliki nilai NA.
Fungsi ini adalah padanan dari `tidyr::drop_na()`.
}
\examples{
# rapi_hapus_na(data)

}
\seealso{
[tidyr::drop_na()]
}
