% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_isi_na}
\alias{rapi_isi_na}
\title{Mengisi nilai hilang (NA) dengan nilai sebelumnya atau berikutnya}
\usage{
rapi_isi_na(data, ..., .direction = "down")
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kolom yang akan diisi.}

\item{.direction}{Arah pengisian: "down", "up", "downup", atau "updown".}
}
\value{
Data frame dengan nilai hilang yang telah diisi.
}
\description{
Mengisi nilai hilang berdasarkan kolom yang dipilih.
Fungsi ini adalah padanan dari `tidyr::fill()`.
}
\examples{
# rapi_isi_na(data, kolom, .direction = "down")

}
\seealso{
[tidyr::fill()]
}
