% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_pisah_kolom}
\alias{rapi_pisah_kolom}
\title{Memisahkan satu kolom menjadi beberapa kolom}
\usage{
rapi_pisah_kolom(data, col, into, sep, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{col}{Kolom yang ingin dipisahkan.}

\item{into}{Nama kolom baru hasil pemisahan.}

\item{sep}{Pola pemisah (misal: "_", "-", atau ekspresi reguler).}

\item{...}{Argumen tambahan untuk `tidyr::separate()`.}
}
\value{
Data frame dengan kolom baru hasil pemisahan.
}
\description{
Memecah kolom berdasarkan pemisah tertentu. Fungsi ini adalah padanan dari `tidyr::separate()`.
}
\examples{
# rapi_pisah_kolom(data, tanggal, into = c("tahun", "bulan"), sep = "-")

}
\seealso{
[tidyr::separate()]
}
