% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_pivot_melebar}
\alias{rapi_pivot_melebar}
\title{Merapikan Data Menjadi Format Melebar}
\usage{
rapi_pivot_melebar(data, names_from, values_from, ...)
}
\arguments{
\item{data}{Data frame yang akan diubah.}

\item{names_from}{Kolom yang berisi nama variabel baru.}

\item{values_from}{Kolom yang berisi nilai variabel baru.}

\item{...}{Argumen tambahan yang diteruskan ke `tidyr::pivot_wider()`.}
}
\value{
Data frame dalam format lebar.
}
\description{
Merapikan dengan mengubah data dari format memanjang ke format lebar.
Fungsi ini adalah padanan dari `tidyr::pivot_wider()`.
}
\examples{
# Contoh:
# rapi_pivot_melebar(data, names_from = bulan, values_from = penjualan)

}
\seealso{
[tidyr::pivot_wider()]
}
