% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{rapi_pivot_memanjang}
\alias{rapi_pivot_memanjang}
\title{Merapikan Data Menjadi Format Memanjang}
\usage{
rapi_pivot_memanjang(data, cols, names_to = "name", values_to = "value", ...)
}
\arguments{
\item{data}{Data frame yang akan diubah.}

\item{cols}{Kolom yang akan diputar menjadi baris.}

\item{names_to}{Nama kolom baru yang menyimpan nama variabel lama.}

\item{values_to}{Nama kolom baru yang menyimpan nilai variabel.}

\item{...}{Argumen tambahan yang diteruskan ke `tidyr::pivot_longer()`.}
}
\value{
Data frame dalam format memanjang.
}
\description{
Merapikan dengan mengubah data dari format lebar ke format memanjang.
Fungsi ini adalah padanan dari `tidyr::pivot_longer()`.
}
\examples{
# Contoh:
# rapi_pivot_memanjang(data, cols = c(jan, feb, mar), names_to = "bulan", values_to = "penjualan")

}
\seealso{
[tidyr::pivot_longer()]
}
