% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\docType{package}
\name{tidyversindonesia}
\alias{tidyversindonesia-package}
\alias{tidyversindonesia}
\title{tidyversindonesia: Jembatan Belajar Tidyverse versi Bahasa Indonesia}
\description{
Package ini menyediakan fungsi-fungsi tidyverse dengan nama dalam Bahasa Indonesia
untuk membantu pembelajar Indonesia memahami konsep data science dengan lebih mudah.
}
\author{
\strong{Maintainer}: Yandi Andiyana \email{yandi.andiyana@gmail.com}

}
