% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_filter}
\alias{transformasi_filter}
\title{Menyaring baris berdasarkan kondisi tertentu}
\usage{
transformasi_filter(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kondisi logika untuk menyaring baris.}
}
\value{
Data frame hasil penyaringan.
}
\description{
Memilih baris yang memenuhi kondisi logika. Fungsi ini adalah padanan dari `dplyr::filter()`.
}
\examples{
# transformasi_filter(data, usia > 30)

}
\seealso{
[dplyr::filter()]
}
