% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_mutasi}
\alias{transformasi_mutasi}
\title{Membuat atau mengubah kolom dalam data}
\usage{
transformasi_mutasi(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Ekspresi untuk membuat atau mengubah kolom.}
}
\value{
Data frame dengan kolom baru atau yang telah diperbarui.
}
\description{
Menambah kolom baru atau memperbarui kolom yang ada.
Fungsi ini adalah padanan dari `dplyr::mutate()`.
}
\examples{
# transformasi_mutasi(data, usia_tahun = usia_bulan / 12)

}
\seealso{
[dplyr::mutate()]
}
