% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_statistik_ringkas}
\alias{transformasi_statistik_ringkas}
\title{Merangkum data menjadi nilai statistik ringkas}
\usage{
transformasi_statistik_ringkas(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble (biasanya sudah dikelompokkan).}

\item{...}{Ekspresi ringkasan yang ingin dihitung.}
}
\value{
Data frame hasil peringkasan.
}
\description{
Menghitung statistik data seperti rata-rata, jumlah, atau nilai maksimum.
Fungsi ini adalah padanan dari `dplyr::summarise()`.
}
\examples{
# transformasi_statistik_ringkas(data, rata_usia = mean(usia, na.rm = TRUE))

}
\seealso{
[dplyr::summarise()]
}
