% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{transformasi_urutkan}
\alias{transformasi_urutkan}
\title{Mengurutkan baris dalam data}
\usage{
transformasi_urutkan(data, ...)
}
\arguments{
\item{data}{Data frame atau tibble.}

\item{...}{Kolom yang digunakan untuk pengurutan.}
}
\value{
Data frame yang telah diurutkan.
}
\description{
Mengurutkan baris berdasarkan satu atau lebih kolom.
Fungsi ini adalah padanan dari `dplyr::arrange()`.
}
\examples{
# transformasi_urutkan(data, desc(usia))

}
\seealso{
[dplyr::arrange()]
}
