% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{visualisasi_ggplot}
\alias{visualisasi_ggplot}
\title{Membuat Plot dari ggplot}
\usage{
visualisasi_ggplot(data, mapping = NULL, ...)
}
\arguments{
\item{data}{Data frame atau tibble yang akan divisualisasikan.}

\item{mapping}{Pemetaan estetika, biasanya dibuat dengan `estetika()`.}

\item{...}{Argumen tambahan untuk `ggplot2::ggplot()`.}
}
\value{
Objek `ggplot` yang dapat ditambahkan layer visualisasi.
}
\description{
Membuat objek grafik utama menggunakan gaya Tidyverse versi Bahasa Indonesia.
Fungsi ini adalah padanan dari `ggplot2::ggplot()`.
}
\examples{
# visualisasi_ggplot(data = penguins, mapping = estetika(x = flipper_length_mm, y = body_mass_g))

}
\seealso{
[ggplot2::ggplot()]
}
