% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversindonesia.R
\name{visualisasi_ggplot_terakhir}
\alias{visualisasi_ggplot_terakhir}
\title{Menampilkan plot ggplot terakhir}
\usage{
visualisasi_ggplot_terakhir()
}
\value{
Objek ggplot dari plot terakhir yang dibuat.
}
\description{
Fungsi ini mengambil plot ggplot yang terakhir dibuat di sesi R saat ini.
Ini adalah padanan dari `ggplot2::last_plot()`.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, hp)) + geom_point()
p
visualisasi_ggplot_terakhir()  # menampilkan plot terakhir
}
}
\seealso{
[ggplot2::last_plot()], [visualisasi_ggplot()]
}
